# swagger/swagger.json
{
	"components": {
		"schemas": {
...
			"malak.Block": {
				"properties": {
					"children": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					},
					"content": {},
					"id": {
						"type": "string"
					},
					"props": {
						"additionalProperties": true,
						"type": "object"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"malak.Update": {
				"properties": {
					"content": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string"
					},
					"created_by": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"is_pinned": {
						"description": "If this update is pinned",
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.UpdateMetadata"
					},
					"reference": {
						"type": "string"
					},
					"sent_at": {
						"type": "string"
					},
					"sent_by": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/malak.UpdateStatus"
					},
					"title": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.UpdateMetadata": {
				"type": "object"
			},
...
			"malak.UpdateStatus": {
				"enum": [
					"draft",
					"sent"
				],
				"type": "string",
				"x-enum-varnames": [
					"UpdateStatusDraft",
					"UpdateStatusSent"
				]
			},
...
			"server.APIStatus": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
...
			"server.contentUpdateRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"update": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"update"
				],
				"type": "object"
			},
...
			"server.createUpdateContent": {
				"properties": {
					"title": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object"
			},
...
			"server.createdUpdateResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"update": {
						"$ref": "#/components/schemas/malak.Update"
					}
				},
				"required": [
					"message",
					"update"
				],
				"type": "object"
			},
...
			"server.fetchUpdateReponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"update": {
						"$ref": "#/components/schemas/malak.Update"
					}
				},
				"required": [
					"message",
					"update"
				],
				"type": "object"
			},
...
			"server.listUpdateResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"meta": {
						"$ref": "#/components/schemas/server.meta"
					},
					"updates": {
						"items": {
							"$ref": "#/components/schemas/malak.Update"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"meta",
					"updates"
				],
				"type": "object"
			},
			"server.meta": {
				"properties": {
					"paging": {
						"$ref": "#/components/schemas/server.pagingInfo"
					}
				},
				"required": [
					"paging"
				],
				"type": "object"
			},
			"server.pagingInfo": {
				"properties": {
					"page": {
						"type": "integer"
					},
					"per_page": {
						"type": "integer"
					},
					"total": {
						"type": "integer"
					}
				},
				"required": [
					"page",
					"per_page",
					"total"
				],
				"type": "object"
			},
...
			"server.sendUpdateRequest": {
				"properties": {
					"emails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"send_at": {
						"type": "integer"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "lanre@ayinke.ventures",
			"name": "Ayinke Ventures"
		},
		"title": "Malak's API documentation",
		"version": "0.1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/workspaces/updates": {
			"get": {
				"description": "List updates",
				"parameters": [
					{
						"description": "Page to query data from. Defaults to 1",
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Number to items to return. Defaults to 10 items",
						"in": "query",
						"name": "per_page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "filter results by the status of the update.",
						"in": "query",
						"name": "status",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listUpdateResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"post": {
				"description": "Create a new update",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.createUpdateContent"
							}
						}
					},
					"description": "update content body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.createdUpdateResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
		"/workspaces/updates/pins": {
			"get": {
				"description": "List pinned updates",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listUpdateResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
		"/workspaces/updates/{reference}": {
			"delete": {
				"description": "Delete a specific update",
				"operationId": "deleteUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"get": {
				"description": "Fetch a specific update",
				"operationId": "fetchUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.fetchUpdateReponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"post": {
				"description": "Send an update to real users",
				"operationId": "sendUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.sendUpdateRequest"
							}
						}
					},
					"description": "request body to send an update",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"put": {
				"description": "Update a specific update",
				"operationId": "updateContent",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.contentUpdateRequest"
							}
						}
					},
					"description": "update content body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5300/v1"
		}
	]
}